Add files via upload
This script will install hotfixes and updates to remote systems in your organization. Additionally it will reboot the servers if there is a requirement prior and after the hotfix install. 
Do the the script requiring CREDSSP it will ask you for your password through the get-credential command 
A list of the prompts are below. 

$Credentials = Get-Credential -UserName "$($env:USERDNSDOMAIN.Substring(0,$($env:USERDNSDOMAIN.indexof("."))))\$env:username" -Message "Enter Credential"
$PackageFileName = Read-Host "Enter the Package File Name to install. Example Exchange2016-KB5000871-x64-en.msp"
$SourceDirectory = Read-Host "Enter the UNC path to the $ExchangeISO file for the install. Example \\ComputerName\NetworkShare"
$TargetDirectory  = Read-Host "Enter the target directory on the Server you wish to install the exchange CU on. Example: D:\Updates"
$PackageOrHotFixID = Read-Host "Enter the KB or Hotfixid. Example KB5000871"